backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-together
langchain-huggingface
sentence-transformers
mysql-connector-python
python-dotenv
